backend/requirements.txt
Flask==3.1.2
flask-cors==4.0.0
SQLAlchemy==2.0.45
psycopg2-binary==2.9.11
requests==2.32.5
python-dotenv==1.0.1
google-genai==1.55.0
huggingface-hub==0.36.0
